requirements.txt
appnope==0.1.0
backcall==0.1.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.3
chardet==4.0.0
Click==7.0
colorama==0.4.4
decorator==4.3.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.10.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.2
idna==2.10
IMDbPY==2020.9.25
ipython==7.18.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.13.1
Jinja2==2.10
lxml==4.6.2
MarkupSafe==1.1.1
omdb==0.10.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.5
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.2.0
requests==2.25.1
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.3.20
traitlets==4.3.2
urllib3==1.26.3
wcwidth==0.1.7
Werkzeug==0.16.0
WTForms==2.2.1
